fix: split out integration tests and fix for update clause
