+++ app/database/queries/api/reset_current_state.sql
@@ -32,8 +32,6 @@ current_versions as (
         where
             input.bucket = s3_object.bucket and
             input.key = s3_object.key
-        order by s3_object.sequencer
-        for update
     ) s3_object
 ),
 -- This selects the single event that is the current state for the key, i.e. the record that represents the current
@@ -42,6 +40,7 @@ current_versions as (
 current_state as (
     select
         s3_object_id,
+        sequencer,
         (
             row_number() over (
                 partition by
@@ -54,10 +53,13 @@ current_state as (
             not current_versions.is_delete_marker
         ) as is_current_state
     from current_versions
-    order by current_versions.sequencer
+),
+-- Apply ordering to prevent deadlocks.
+current_state_ordered as (
+    select * from current_state order by current_state.sequencer for update
 )
 update s3_object
-set is_current_state = current_state.is_current_state
-from current_state
-where s3_object.s3_object_id = current_state.s3_object_id
+set is_current_state = current_state_ordered.is_current_state
+from current_state_ordered
+where s3_object.s3_object_id = current_state_ordered.s3_object_id
 returning s3_object.s3_object_id, s3_object.is_current_state;
